A simple attempt at send cancellation
This PR attempts to implement the capability of cancelling a pending or
in-flight transfer. The method is to start a goroutine during the send
process that selects on a context's Done() channel; if signalled, the
goroutine calls Close() on the transportCryptor. This has the effect of
cancelling a send operation. The function ReadPake() is also extended to
take a context argument, and will select on the context's Done() channel
at the same time it waits for a client message. This seems to serve my
immediate purposes, but it may not be complete. I'm happy to try to
extend it to cover cancellation in more situations!